Add serialization, move apps to separate folder

+++ commons/common-domain/src/main/kotlin/com/surfshop/common/Speed.kt
@@ -1,5 +1,5 @@
 package com.surfshop.common
 
-data class Speed(val beaufort: Int) {
+data class Speed(val beaufort: Long) {
 
 }
